Add time delay
Adding delay because of the pn532 module. If you keep your card on the pn532 module, it keeps recognizing the card and will return the card uid a lot in a little amount of time, which will make it super busy. I did not want that to happen, so I added some delay.